organize utility functions for sims a bit better
